chore(zero-cache): move MessageProcessor out of IncrementalSyncer
